Tune app_booter

+++ app_booter/source/string.h
@@ -1,9 +1,9 @@
 #ifndef STRING_H_
 #define STRING_H_
 
-#include <stdio.h>
+#include <stdint.h>
 
 void *memcpy(void *dst, const void *src, size_t len);
-void *memset(void *b, int c, size_t len);
+void *memset(void *b, uint8_t c, size_t len);
 
 #endif
